// web/projects/em/src/app/settings/schedule/model/reports-options.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
export enum ReportOptions {
   PDF = "PDF",
   POWERPOINT = "POWERPOINT",
   EXCEL = "EXCEL",
   EXCEL_SHEET = "EXCEL_SHEET",
   EXCEL_DATA = "EXCEL_DATA",
   RTF = "RTF",
   RTF_LAYOUT = "RTF_LAYOUT",
   HTML = "HTML",
   HTML_BUNDLE = "HTML_BUNDLE",
   HTML_NO_PAGINATION= "HTML_NO_PAGINATION",
   HTML_BUNDLE_NO_PAGINATION = "HTML_BUNDLE_NO_PAGINATION",
   HTML_NO_PAGINATION_EMAIL = "HTML_NO_PAGINATION_EMAIL",
   TEXT = "TEXT",
   CSV = "CSV",
   REPORT = "REPORT",
   SVG = "SVG"
}